Frontend/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
authlib>=1.3.1
httpx>=0.27.0
itsdangerous>=2.2.0
python-dotenv>=1.0.1
pandas>=2.2.0
groq>=0.9.0
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
tabulate>=0.9.0
numpy>=1.26.0
